Add files via upload
This project provides a detailed analysis of the University of North Carolina’s field hockey team's offensive tendencies, with a specific focus on their strategies in the attacking circle and shooting patterns over the last four games. The purpose of this analysis was to identify the elements that make UNC’s attack so effective and to offer actionable insights to support game preparation. By examining their movement patterns, positioning, and preferred methods for creating scoring opportunities, this work sheds light on how UNC consistently puts pressure on opposing defenses. Additionally, the analysis includes an in-depth look at shooting tendencies, such as shot frequency, selection, and accuracy, to understand when and how they are most effective. This comprehensive review was shared with the coaching staff and team to inform strategic planning ahead of the game.